+++ src/main/kotlin/com/github/yoannteruel/jetbrainsworktreeplugin/actions/OpenWorktreeAction.kt
@@ -1,11 +1,9 @@
 package com.github.yoannteruel.jetbrainsworktreeplugin.actions
 
-import com.github.yoannteruel.jetbrainsworktreeplugin.settings.WorktreeSettingsService
 import com.github.yoannteruel.jetbrainsworktreeplugin.ui.WorktreeToolWindowPanel
 import com.intellij.ide.impl.ProjectUtil
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.components.service
 import java.nio.file.Path
 
 class OpenWorktreeAction : AnAction() {
@@ -15,12 +13,7 @@ class OpenWorktreeAction : AnAction() {
         val worktree = e.getData(WorktreeToolWindowPanel.SELECTED_WORKTREE) ?: return
         val worktreePath = Path.of(worktree.path)
 
-        val settings = project.service<WorktreeSettingsService>()
-        if (settings.state.openWorktreeInNewWindow) {
-            ProjectUtil.openOrImport(worktreePath)
-        } else {
-            ProjectUtil.openOrImport(worktreePath, project, false)
-        }
+        ProjectUtil.openOrImport(worktreePath, project, true)
     }
 
     override fun update(e: AnActionEvent) {